pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fast-readme-ai"
version = "1.0.0"
description = "AI-powered README generator for developers"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["readme", "documentation", "ai", "gemini", "cli", "developer-tools"]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "gitpython>=3.1.40",
    "streamlit>=1.35.0",
    "requests>=2.31.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
fast-readme = "cli.main:app"

[tool.setuptools]
py-modules = ["config"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "cli*", "api*"]
